Refactor: 파일명, 확장자 변경 및 error 처리
- jotai.js 파일명 변경
- ChartSection error 처리
- useLazyImageObserver.js 확장자 변경
