+++ src/components/Modal/Modal.module.scss
@@ -223,6 +223,12 @@
     background-color: var(--bg-modal);
   }
 
+  .error {
+    padding: 120px 0 0;
+    text-align: center;
+    font-size: 1.6rem;
+  }
+
   .main {
     margin-top: 68px;
   }
