+++ src/pages/List/List.module.scss
@@ -130,6 +130,11 @@
         width: 100%;
       }
     }
+    .error {
+      padding: 40px 0;
+      text-align: center;
+      font-size: 1.6rem;
+    }
     .chartList {
       display: flex;
       gap: 17px;
