+++ src/pages/MyPage/MyPage.module.scss
@@ -64,14 +64,6 @@
             top: -10px;
             left: 60px;
           }
-          .deleteIcon {
-            width: 22px;
-            height: 22px;
-            @media screen and (min-width: $tablet) {
-              width: 30px;
-              height: 30px;
-            }
-          }
         }
         .roundImage {
           width: 70px;
